Form Validation
Form now shoes if data entered is empty and if the 2 passwords entered are the same or different. Note that if a different password is entered on the second input the error is different. The Functions are in place if you want to add more regular expressions to the form validation. Email now correctly checks if the email format is correct.